# schedule/schedule.go
package schedule

import (
	"bytes"
	"context"
	"encoding/json"
	"strings"
	"time"

	"github.com/spy16/clockwork"
	"github.com/spy16/clockwork/pkg/cron"
)

const (
	StatusDone     = "DONE"
	StatusActive   = "ACTIVE"
	StatusDisabled = "DISABLED"
)

//go:generate mockery --with-expecter --keeptree --case snake --name Scheduler
//go:generate mockery --with-expecter --keeptree --case snake --name ChangeLogger

// ChangeLogger is responsible for publishing changes that occur in schedules.
type ChangeLogger interface {
	// Publish should publish the schedule with the correct change type.
	// changeType can be 1 - Created, 2 - Updated and 3 - Deleted.
	Publish(ctx context.Context, changeType int, sc Schedule) error
}

// Scheduler is responsible for managing schedule definitions and creating
// execution requests as per the schedules.
type Scheduler interface {
	// List returns the paginated list of schedules using the offset and
	// count. If listing is not supported by the Scheduler backend, this
	// may return ErrUnsupported.
	List(ctx context.Context, offset, count int) ([]Schedule, error)

	// Get returns schedule definition with given id. Returns ErrNotFound
	// if no such schedule exists.
	Get(ctx context.Context, scheduleID string) (*Schedule, error)

	// Put validates and creates/updates schedule in the Scheduler. This will
	// also create the execution points as required by the definition.
	Put(ctx context.Context, sc Schedule, isUpdate bool, req ...Execution) error

	// Del removes the schedule with given id and all its execution points
	// from the Scheduler. Returns ErrNotFound if schedule does not exist.
	Del(ctx context.Context, scheduleID string) error

	// Run runs the process required to identify schedules ready for execution
	// and invokes onReady for each of them. Run blocks until a critical error
	// occurs or the context is cancelled.
	Run(ctx context.Context, onReady OnReadyFunc) error
}

// OnReadyFunc is invoked by the Scheduler whenever a schedule request is ready
// for execution.
type OnReadyFunc func(ctx context.Context, sc Schedule, current Execution) (next *Execution, err error)

// Execution represents execution request for a schedule. Such requests are
// generated by the Scheduler and delivered to the Service.
type Execution struct {
	Manual     bool      `json:"manual"`
	Version    int64     `json:"version"`
	EnqueueAt  time.Time `json:"enqueue_at"`
	ScheduleID string    `json:"schedule_id"`
}

// Schedule represents a schedule that needs to be executed at some specific
// points in time.
type Schedule struct {
	ID              string    `json:"id"`
	Tags            []string  `json:"tags"`
	Status          string    `json:"status"`
	Crontab         string    `json:"crontab"`
	Triggers        []int64   `json:"triggers,omitempty"`
	Version         int64     `json:"version"`
	Category        string    `json:"category"`
	Payload         string    `json:"payload"`
	ClientID        string    `json:"client_id"`
	CreatedAt       time.Time `json:"created_at"`
	UpdatedAt       time.Time `json:"updated_at"`
	EnqueueCount    int       `json:"enqueue_count"` // number of times this schedule has been enqueued (not executed).
	NextExecutionAt time.Time `json:"last_enqueued_at"`
}

// Updates can be applied on a Schedule to modify it as required.
type Updates struct {
	Disable  *bool    `json:"disable"`
	Crontab  string   `json:"crontab"`
	Category string   `json:"category"`
	Trigger  int64    `json:"trigger"`
	Tags     []string `json:"tags"`
	Payload  *string  `json:"payload"`
}

// ComputeNext returns the next point of execution for the schedule relative to the
// given timestamp, computed as per the crontab specified.
func (sc *Schedule) ComputeNext(relativeTo time.Time) (time.Time, error) {
	schedule, err := cron.Parse(sc.Crontab)
	if err != nil {
		return time.Time{}, clockwork.ErrInvalid.WithCausef("invalid crontab")
	}
	return schedule.Next(relativeTo), nil
}

// Validate ensures the Schedule definition is valid by sanitising and applying
// relevant checks. Returns ErrInvalid if any check fails.
func (sc *Schedule) Validate() error {
	sc.ID = strings.TrimSpace(sc.ID)
	sc.ClientID = strings.TrimSpace(sc.ClientID)
	sc.Crontab = strings.TrimSpace(sc.Crontab)
	if sc.Status == "" {
		sc.Status = StatusActive
	}

	switch {
	case sc.ID == "":
		return clockwork.ErrInvalid.WithCausef("id must not be empty")

	case sc.Crontab == "":
		return clockwork.ErrInvalid.WithCausef("crontab must be specified")

	case sc.ClientID == "":
		return clockwork.ErrInvalid.WithCausef("client_id must not be empty")

	case sc.Category == "":
		return clockwork.ErrInvalid.WithCausef("category must not be empty")
	}

	_, err := cron.Parse(sc.Crontab)
	if err != nil {
		return clockwork.ErrInvalid.WithCausef("invalid crontab").WithMsgf("%s", err.Error())
	}

	return nil
}

// Apply applies the updates to the schedule. Returns true if a reschedule
// is needed due to the changes done.
func (sc *Schedule) Apply(updates Updates, clockFn ClockFunc) (rescheduleNeeded bool) {
	defer func() { sc.UpdatedAt = clockFn() }()

	if updates.Tags != nil {
		sc.Tags = updates.Tags
	}

	if updates.Payload != nil && *updates.Payload != sc.Payload {
		sc.Payload = *updates.Payload
	}

	updates.Category = strings.TrimSpace(updates.Category)
	if updates.Category != "" && updates.Category != sc.Category {
		sc.Category = updates.Category
	}

	incrVersion := false // only increments when scheduling timeline changes.

	updates.Crontab = strings.TrimSpace(updates.Crontab)
	if updates.Crontab != "" && updates.Crontab != sc.Crontab {
		incrVersion = true // crontab update affects scheduling timeline.
		sc.Crontab = updates.Crontab
	}

	if updates.Disable != nil && sc.Status != StatusDone {
		currentlyDisabled := sc.Status == StatusDisabled
		incrVersion = *updates.Disable != currentlyDisabled

		if *updates.Disable {
			sc.Status = StatusDisabled
		} else {
			sc.Status = StatusActive
		}
	}

	if incrVersion {
		// Incrementing version automatically invalidates all previously inserted execution
		// points on the scheduler. So, in case of any timeline changes, we need to increment
		// the version and recompute the execution points.
		sc.Version++
	}

	if updates.Trigger > 0 {
		sc.Status = StatusActive
		return true
	}

	return incrVersion && sc.Status != StatusDisabled
}

// JSON returns the json marshaled version of the job definition.
func (sc *Schedule) JSON() []byte {
	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(sc); err != nil {
		// note: should not happen.
		panic(err)
	}
	return buf.Bytes()
}